+++ test/napkinsketch_book/api_reference_generated_test.clj
@@ -240,7 +240,9 @@
   wave
   {:x (range 30),
    :y
-   (map (fn* [p1__79706#] (Math/sin (* p1__79706# 0.3))) (range 30))}))
+   (map
+    (fn* [p1__181166#] (Math/sin (* p1__181166# 0.3)))
+    (range 30))}))
 
 
 (def v47_l159 (-> wave (sk/lay-line :x :y)))
@@ -359,9 +361,9 @@
     :y
     (map
      (fn*
-      [p1__79707#]
+      [p1__181167#]
       (+
-       (Math/sin (* p1__79707# 0.2))
+       (Math/sin (* p1__181167# 0.2))
        (* 0.3 (- (rng/drandom r) 0.5))))
      (range 50))})))