+++ test/napkinsketch_book/cookbook_generated_test.clj
@@ -102,8 +102,8 @@
    :value
    (map
     (fn*
-     [p1__79384#]
-     (+ 100.0 (* 30.0 (Math/sin (* (double p1__79384#) 0.12)))))
+     [p1__180741#]
+     (+ 100.0 (* 30.0 (Math/sin (* (double p1__180741#) 0.12)))))
     (range 52))}))
 
 
@@ -216,7 +216,25 @@
 
 
 (def
- v35_l162
+ v35_l163
+ (->
+  (rdatasets/datasets-iris)
+  (sk/lay-point :sepal-length :sepal-width {:alpha 0.3})
+  (sk/lay-point
+   {:data {:sepal-length [5.0 6.5], :sepal-width [3.5 3.0]},
+    :x :sepal-length,
+    :y :sepal-width,
+    :color "red",
+    :size 6})))
+
+
+(deftest
+ t36_l170
+ (is ((fn [v] (= 152 (:points (sk/svg-summary v)))) v35_l163)))
+
+
+(def
+ v38_l177
  (def
   experiment
   {:condition ["A" "B" "C" "D"],
@@ -226,7 +244,7 @@
 
 
 (def
- v36_l168
+ v39_l183
  (->
   experiment
   (sk/lay-point :condition :mean {:size 5})
@@ -234,18 +252,18 @@
 
 
 (deftest
- t37_l172
+ t40_l187
  (is
   ((fn
     [v]
     (let
      [s (sk/svg-summary v)]
      (and (= 4 (:points s)) (= 12 (:lines s)))))
-   v36_l168)))
+   v39_l183)))
 
 
 (def
- v39_l180
+ v42_l195
  (->
   experiment
   (sk/lay-lollipop :condition :mean)
@@ -253,18 +271,18 @@
 
 
 (deftest
- t40_l184
+ t43_l199
  (is
   ((fn
     [v]
     (let
      [s (sk/svg-summary v)]
      (and (= 4 (:points s)) (= 16 (:lines s)))))
-   v39_l180)))
+   v42_l195)))
 
 
 (def
- v42_l192
+ v45_l207
  (->
   (rdatasets/datasets-iris)
   (sk/lay-point :species :sepal-length {:alpha 0.3, :jitter 5})
@@ -272,18 +290,18 @@
 
 
 (deftest
- t43_l196
+ t46_l211
  (is
   ((fn
     [v]
     (let
      [s (sk/svg-summary v)]
      (and (= 153 (:points s)) (= 3 (:lines s)))))
-   v42_l192)))
+   v45_l207)))
 
 
 (def
- v45_l204
+ v48_l219
  (->
   (rdatasets/reshape2-tips)
   (sk/view :total-bill :tip {:color :smoker})
@@ -296,7 +314,7 @@
 
 
 (deftest
- t46_l212
+ t49_l227
  (is
   ((fn
     [v]
@@ -306,11 +324,11 @@
       (pos? (:points s))
       (= 2 (:lines s))
       (some #{"Tipping Behavior"} (:texts s)))))
-   v45_l204)))
+   v48_l219)))
 
 
 (def
- v48_l224
+ v51_l239
  (->
   (rdatasets/datasets-iris)
   (sk/view :sepal-length :sepal-width {:color :species})
@@ -320,18 +338,18 @@
 
 
 (deftest
- t49_l230
+ t52_l245
  (is
   ((fn
     [v]
     (let
      [s (sk/svg-summary v)]
      (and (= 1 (:panels s)) (pos? (:points s)) (pos? (:lines s)))))
-   v48_l224)))
+   v51_l239)))
 
 
 (def
- v51_l239
+ v54_l254
  (->
   (rdatasets/reshape2-tips)
   (sk/lay-bar :day {:color :sex})
@@ -339,12 +357,12 @@
 
 
 (deftest
- t52_l243
- (is ((fn [v] (pos? (:polygons (sk/svg-summary v)))) v51_l239)))
+ t55_l258
+ (is ((fn [v] (pos? (:polygons (sk/svg-summary v)))) v54_l254)))
 
 
 (def
- v53_l245
+ v56_l260
  (->
   (rdatasets/reshape2-tips)
   (sk/lay-stacked-bar :day {:color :sex})
@@ -352,12 +370,12 @@
 
 
 (deftest
- t54_l249
- (is ((fn [v] (pos? (:polygons (sk/svg-summary v)))) v53_l245)))
+ t57_l264
+ (is ((fn [v] (pos? (:polygons (sk/svg-summary v)))) v56_l260)))
 
 
 (def
- v56_l256
+ v59_l271
  (def
   daily-temps
   {:day (range 1 15),
@@ -365,7 +383,7 @@
 
 
 (def
- v57_l260
+ v60_l275
  (->
   daily-temps
   (sk/lay-step :day :temp {:color "#2196F3"})
@@ -374,18 +392,18 @@
 
 
 (deftest
- t58_l265
+ t61_l280
  (is
   ((fn
     [v]
     (let
      [s (sk/svg-summary v)]
      (and (pos? (:lines s)) (pos? (:points s)))))
-   v57_l260)))
+   v60_l275)))
 
 
 (def
- v60_l274
+ v63_l289
  (->
   (rdatasets/datasets-iris)
   (sk/lay-point
@@ -396,18 +414,18 @@
 
 
 (deftest
- t61_l278
+ t64_l293
  (is
   ((fn
     [v]
     (let
      [s (sk/svg-summary v)]
      (and (pos? (:points s)) (pos? (:lines s)))))
-   v60_l274)))
+   v63_l289)))
 
 
 (def
- v63_l286
+ v66_l301
  (def
   top5
   (->
@@ -417,7 +435,7 @@
 
 
 (def
- v64_l288
+ v67_l303
  (->
   top5
   (sk/lay-point :sepal-length :sepal-width {:size 5})
@@ -425,7 +443,7 @@
 
 
 (deftest
- t65_l292
+ t68_l307
  (is
   ((fn
     [v]
@@ -434,13 +452,13 @@
      (and
       (pos? (:points s))
       (some
-       (fn* [p1__79385#] (= "virginica" p1__79385#))
+       (fn* [p1__180742#] (= "virginica" p1__180742#))
        (:texts s)))))
-   v64_l288)))
+   v67_l303)))
 
 
 (def
- v67_l300
+ v70_l315
  (->
   (rdatasets/datasets-iris)
   (sk/lay-point :sepal-length :sepal-width {:color :species})
@@ -451,18 +469,18 @@
 
 
 (deftest
- t68_l307
+ t71_l322
  (is
   ((fn
     [v]
     (let
      [s (sk/svg-summary v)]
      (and (= 1 (:panels s)) (pos? (:points s)))))
-   v67_l300)))
+   v70_l315)))
 
 
 (def
- v70_l316
+ v73_l331
  (->
   (rdatasets/datasets-iris)
   (sk/view :sepal-length :sepal-width {:color :species})
@@ -473,23 +491,23 @@
 
 
 (deftest
- t71_l323
+ t74_l338
  (is
   ((fn
     [v]
     (let
      [s (sk/svg-summary v)]
      (and (pos? (:points s)) (= 3 (:lines s)))))
-   v70_l316)))
+   v73_l331)))
 
 
 (def
- v73_l332
+ v76_l347
  (->
   {:x (range 20),
    :y
-   (map (fn* [p1__79386#] (Math/sin (/ p1__79386# 3.0))) (range 20)),
-   :change (map (fn* [p1__79387#] (- p1__79387# 10)) (range 20))}
+   (map (fn* [p1__180743#] (Math/sin (/ p1__180743# 3.0))) (range 20)),
+   :change (map (fn* [p1__180744#] (- p1__180744# 10)) (range 20))}
   (sk/lay-point :x :y {:color :change})
   (sk/options
    {:color-scale :diverging,
@@ -498,18 +516,18 @@
 
 
 (deftest
- t74_l340
+ t77_l355
  (is
   ((fn
     [v]
     (let
      [s (sk/svg-summary v)]
      (and (= 1 (:panels s)) (= 20 (:points s)))))
-   v73_l332)))
+   v76_l347)))
 
 
 (def
- v76_l348
+ v79_l363
  (->
   (rdatasets/datasets-iris)
   (sk/view :sepal-length :sepal-width {:color :species})
@@ -519,18 +537,18 @@
 
 
 (deftest
- t77_l354
+ t80_l369
  (is
   ((fn
     [v]
     (let
      [s (sk/svg-summary v)]
      (and (= 150 (:points s)) (= 3 (:lines s)) (= 3 (:polygons s)))))
-   v76_l348)))
+   v79_l363)))
 
 
 (def
- v79_l363
+ v82_l378
  (def
   iris-sepal
   (->
@@ -540,7 +558,7 @@
 
 
 (def
- v80_l368
+ v83_l383
  (def
   iris-petal
   (->
@@ -550,23 +568,23 @@
 
 
 (def
- v81_l373
+ v84_l388
  (sk/arrange
   [iris-sepal iris-petal]
   {:title "Iris Dashboard", :cols 2}))
 
 
 (deftest
- t82_l376
+ t85_l391
  (is
   ((fn
     [v]
     (and (= :div (first v)) (= :kind/hiccup (:kindly/kind (meta v)))))
-   v81_l373)))
+   v84_l388)))
 
 
 (def
- v84_l383
+ v87_l398
  (def
   top-cities
   {:city ["Tokyo" "Delhi" "Shanghai" "São Paulo" "Mumbai"],
@@ -575,7 +593,7 @@
 
 
 (def
- v85_l388
+ v88_l403
  (->
   top-cities
   (sk/lay-point :area :population)
@@ -584,7 +602,7 @@
 
 
 (deftest
- t86_l393
+ t89_l408
  (is
   ((fn
     [v]
@@ -593,11 +611,11 @@
      (and
       (= 5 (:points s))
       (every? (set (:texts s)) ["Tokyo" "Delhi"]))))
-   v85_l388)))
+   v88_l403)))
 
 
 (def
- v88_l401
+ v91_l416
  (let
   [r
    (rng/rng :jdk 77)
@@ -606,8 +624,8 @@
    ys
    (map
     (fn*
-     [p1__79388#]
-     (+ (* 3 p1__79388#) 5 (* 2 (- (rng/drandom r) 0.5))))
+     [p1__180745#]
+     (+ (* 3 p1__180745#) 5 (* 2 (- (rng/drandom r) 0.5))))
     xs)]
   (->
    {:x xs, :y ys}
@@ -617,7 +635,7 @@
 
 
 (deftest
- t89_l412
+ t92_l427
  (is
   ((fn
     [v]
@@ -627,11 +645,11 @@
       (= 20 (:points s))
       (= 1 (:lines s))
       (some #{"Simulated: y = 3x + 5 + noise"} (:texts s)))))
-   v88_l401)))
+   v91_l416)))
 
 
 (def
- v91_l423
+ v94_l438
  (->
   (rdatasets/palmerpenguins-penguins)
   (sk/lay-point :bill-length-mm :bill-depth-mm {:color :species})
@@ -639,18 +657,18 @@
 
 
 (deftest
- t92_l427
+ t95_l442
  (is
   ((fn
     [v]
     (let
      [s (sk/svg-summary v)]
      (and (= 1 (:panels s)) (= 342 (:points s)))))
-   v91_l423)))
+   v94_l438)))
 
 
 (def
- v94_l433
+ v97_l448
  (->
   (rdatasets/palmerpenguins-penguins)
   (sk/view :bill-length-mm :bill-depth-mm {:color :species})
@@ -660,18 +678,18 @@
 
 
 (deftest
- t95_l439
+ t98_l454
  (is
   ((fn
     [v]
     (let
      [s (sk/svg-summary v)]
      (and (= 342 (:points s)) (= 3 (:lines s)))))
-   v94_l433)))
+   v97_l448)))
 
 
 (def
- v97_l446
+ v100_l461
  (->
   (rdatasets/palmerpenguins-penguins)
   (sk/lay-point :bill-length-mm :bill-depth-mm {:color :species})
@@ -681,18 +699,18 @@
 
 
 (deftest
- t98_l451
+ t101_l466
  (is
   ((fn
     [v]
     (let
      [s (sk/svg-summary v)]
      (and (= 342 (:points s)) (= 1 (:lines s)))))
-   v97_l446)))
+   v100_l461)))
 
 
 (def
- v100_l457
+ v103_l472
  (->
   (rdatasets/palmerpenguins-penguins)
   (sk/lay-bar :island {:color :species})
@@ -700,18 +718,18 @@
 
 
 (deftest
- t101_l461
+ t104_l476
  (is
   ((fn
     [v]
     (let
      [s (sk/svg-summary v)]
      (and (= 1 (:panels s)) (pos? (:polygons s)))))
-   v100_l457)))
+   v103_l472)))
 
 
 (def
- v103_l467
+ v106_l482
  (->
   (rdatasets/palmerpenguins-penguins)
   (sk/view :flipper-length-mm :body-mass-g {:color :species})
@@ -721,18 +739,18 @@
 
 
 (deftest
- t104_l473
+ t107_l488
  (is
   ((fn
     [v]
     (let
      [s (sk/svg-summary v)]
      (and (= 342 (:points s)) (= 3 (:lines s)))))
-   v103_l467)))
+   v106_l482)))
 
 
 (def
- v106_l479
+ v109_l494
  (->
   (rdatasets/palmerpenguins-penguins)
   (sk/lay-histogram :body-mass-g {:color :species})
@@ -740,18 +758,18 @@
 
 
 (deftest
- t107_l483
+ t110_l498
  (is
   ((fn
     [v]
     (let
      [s (sk/svg-summary v)]
      (and (= 1 (:panels s)) (pos? (:polygons s)))))
-   v106_l479)))
+   v109_l494)))
 
 
 (def
- v109_l491
+ v112_l506
  (->
   (rdatasets/reshape2-tips)
   (sk/view :total-bill :tip {:color :smoker})
@@ -764,18 +782,18 @@
 
 
 (deftest
- t110_l498
+ t113_l513
  (is
   ((fn
     [v]
     (let
      [s (sk/svg-summary v)]
      (and (= 244 (:points s)) (= 2 (:lines s)))))
-   v109_l491)))
+   v112_l506)))
 
 
 (def
- v112_l504
+ v115_l519
  (->
   (rdatasets/reshape2-tips)
   (sk/lay-bar :day {:color :time})
@@ -783,18 +801,18 @@
 
 
 (deftest
- t113_l508
+ t116_l523
  (is
   ((fn
     [v]
     (let
      [s (sk/svg-summary v)]
      (and (= 1 (:panels s)) (pos? (:polygons s)))))
-   v112_l504)))
+   v115_l519)))
 
 
 (def
- v115_l514
+ v118_l529
  (->
   (rdatasets/reshape2-tips)
   (sk/lay-stacked-bar :day {:color :time})
@@ -802,18 +820,18 @@
 
 
 (deftest
- t116_l518
+ t119_l533
  (is
   ((fn
     [v]
     (let
      [s (sk/svg-summary v)]
      (and (= 1 (:panels s)) (pos? (:polygons s)))))
-   v115_l514)))
+   v118_l529)))
 
 
 (def
- v118_l524
+ v121_l539
  (->
   (rdatasets/reshape2-tips)
   (sk/lay-bar :day {:color :sex})
@@ -822,18 +840,18 @@
 
 
 (deftest
- t119_l529
+ t122_l544
  (is
   ((fn
     [v]
     (let
      [s (sk/svg-summary v)]
      (and (= 1 (:panels s)) (pos? (:polygons s)))))
-   v118_l524)))
+   v121_l539)))
 
 
 (def
- v121_l537
+ v124_l552
  (->
   (rdatasets/ggplot2-mpg)
   (sk/view :displ :hwy {:color :class})
@@ -843,18 +861,18 @@
 
 
 (deftest
- t122_l543
+ t125_l558
  (is
   ((fn
     [v]
     (let
      [s (sk/svg-summary v)]
      (and (= 234 (:points s)) (pos? (:lines s)))))
-   v121_l537)))
+   v124_l552)))
 
 
 (def
- v124_l549
+ v127_l564
  (->
   (rdatasets/ggplot2-mpg)
   (sk/lay-point :displ :cty {:color :drv})
@@ -862,18 +880,18 @@
 
 
 (deftest
- t125_l553
+ t128_l568
  (is
   ((fn
     [v]
     (let
      [s (sk/svg-summary v)]
      (and (= 1 (:panels s)) (= 234 (:points s)))))
-   v124_l549)))
+   v127_l564)))
 
 
 (def
- v127_l559
+ v130_l574
  (->
   (rdatasets/ggplot2-mpg)
   (sk/lay-bar :drv)
@@ -881,18 +899,18 @@
 
 
 (deftest
- t128_l563
+ t131_l578
  (is
   ((fn
     [v]
     (let
      [s (sk/svg-summary v)]
      (and (= 1 (:panels s)) (pos? (:polygons s)))))
-   v127_l559)))
+   v130_l574)))
 
 
 (def
- v130_l583
+ v133_l598
  (->
   (rdatasets/ggplot2-diamonds)
   (tc/head 500)
@@ -901,12 +919,12 @@
 
 
 (deftest
- t131_l588
- (is ((fn [v] (= 500 (:points (sk/svg-summary v)))) v130_l583)))
+ t134_l603
+ (is ((fn [v] (= 500 (:points (sk/svg-summary v)))) v133_l598)))
 
 
 (def
- v133_l595
+ v136_l610
  (->
   (rdatasets/ggplot2-diamonds)
   (sk/lay-point :carat :price {:color :cut, :alpha 0.3})
@@ -914,4 +932,4 @@
    {:title "Diamonds (53,940 rows, BufferedImage)", :format :bufimg})))
 
 
-(deftest t134_l600 (is ((fn [v] (some? v)) v133_l595)))
+(deftest t137_l615 (is ((fn [v] (some? v)) v136_l610)))